rename eventstore
