+++ eventstore/eventstore/db_client.py
@@ -1,4 +1,4 @@
-from app.event_store import EventStore
+from eventstore.event_store import EventStore
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 from event_schema.EventSchema import Event
